feat: add scripted layout
